requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
coverage==7.13.4
cryptography==46.0.5
deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.8.0
mako==1.3.10
markupsafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
redis==7.3.0
rsa==4.9.1
ruff==0.15.6
six==1.17.0
slowapi==0.1.9
sqlalchemy==2.0.48
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
wrapt==2.1.2
